Add /a*?/ -> /a+?/ and /a*+/ -> /a++/ mutations
- Adds quantifier restriction mutation for reluctant zero or more.
- Adds quantifier restriction mutation for possive zero or more.
- Renames GreedyZeroOrMore to simply ZeroOrMore
