+++ config/flay.yml
@@ -1,3 +1,3 @@
 ---
 threshold: 16
-total_score: 1301
+total_score: 1327
