+++ lib/mutest.rb
@@ -96,7 +96,7 @@ require 'mutest/mutator/node/regexp/capture_group'
 require 'mutest/mutator/node/regexp/character_type'
 require 'mutest/mutator/node/regexp/end_of_line_anchor'
 require 'mutest/mutator/node/regexp/end_of_string_or_before_end_of_line_anchor'
-require 'mutest/mutator/node/regexp/greedy_zero_or_more'
+require 'mutest/mutator/node/regexp/zero_or_more'
 require 'mutest/mutator/node/literal'
 require 'mutest/mutator/node/literal/boolean'
 require 'mutest/mutator/node/literal/range'
