--- meta/regexp/regexp_greedy_zero_or_more.rb
@@ -1,10 +1,0 @@
-Mutest::Meta::Example.add :regexp_greedy_zero_or_more do
-  source '/\d*/'
-
-  singleton_mutations
-  regexp_mutations
-
-  mutation '/\d/'
-  mutation '/\d+/'
-  mutation '/\D*/'
-end
